Added transfer delete and moved func
